\name{data_classic3}
\docType{data}
\alias{data_classic3}
\title{Document-term matrix of three corpora with sparse format}
\description{
 This dataset, with dimensions 3891 x 4303, contains three medical corpora, CISI, CRAN and MED. This data object is compounded of the matrix A_ijx with the format Coordinate List (row, column, value), plus un vector with the labels for the rows, and a char array with the name of the dataset.
}
\keyword{datasets}
