\name{data_cstr}
\docType{data}
\alias{data_cstr}
\title{Document-term matrix of four research topics with sparse format}
\description{
 This dataset, with dimensions 475 x 1000, contains four research topics Natural Language Processing, Robotics/Vision, Systems, and Theory. This data object is compounded of the matrix A_ijx with the format Coordinate List (row, column, value), plus un vector with the labels for the rows, and a char array with the name of the dataset.
}
\keyword{datasets}
