\name{data_news4}
\docType{data}
\alias{data_news4}
\title{Document-term matrix of four newgroups with sparse format}
\description{
 This dataset, with dimensions 372 x 100, contains four newsgroups asci.crypt, asci.space, asci.med, and asoc.religion.christian. This data object is compounded of the matrix A_ijx with the format Coordinate List (row, column, value), plus un vector with the labels for the rows, and a char array with the name of the dataset.
}
\keyword{datasets}
