\name{data_ng20}
\docType{data}
\alias{data_ng20}
\title{Document-term matrix of 20 newgroups with sparse format}
\description{
  This dataset, with dimensions 19949 x 43586, contains twenty newsgroups. This data object is compounded of the matrix A_ijx with the format Coordinate List (row, column, value), plus un vector with the labels for the rows, and a char array with the name of the dataset.
}
\keyword{datasets}
