\name{data_ng5}
\docType{data}
\alias{data_ng5}
\title{Document-term matrix of five newgroups with sparse format}
\description{
  This dataset, with dimensions 499 x 2000, contains five newsgroups comp.os.ms-windows, comp. windows.x, rec.motorcycles, and sci.crypt, sci.space. This data object is compounded of the matrix A_ijx with the format Coordinate List (row, column, value), plus un vector with the labels for the rows, and a char array with the name of the dataset.
}
\keyword{datasets}
