\name{data_webace}
\docType{data}
\alias{data_webace}
\title{Document-term matrix of 20 topics with sparse format}
\description{
 This dataset, with dimensions 2340 x 1000, contains news articles from 20 different topics. This data object is compounded of the matrix A_ijx with the format Coordinate List (row, column, value), plus un vector with the labels for the rows, and a char array with the name of the dataset.
}
\keyword{datasets}
